saluki_error

Type Alias GenericError

pub type GenericError = Error;

A wrapper around a dynamic error type.

GenericError works a lot like Box<dyn std::error::Error>, but with these differences:

GenericError requires that the error is Send, Sync, and 'static.

GenericError guarantees that a backtrace is available, even if the underlying error type does not provide one.

GenericError is represented as a narrow pointer — exactly one word in size instead of two.

Aliased Type

pub struct GenericError { /* private fields */ }